requirements.txt
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
bidict==0.22.1
click==8.1.3
fastapi==0.70.0
fastapi-socketio==0.0.10
FastAPI-SQLAlchemy==0.2.1
h11==0.14.0
hiredis==2.2.1
httptools==0.5.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.6
pydantic==1.10.4
PyJWT==2.2.0
python-dotenv==0.19.0
python-engineio==4.4.1
python-socketio==5.8.0
PyYAML==6.0
redis==4.4.2
sniffio==1.3.0
SQLAlchemy==1.4.26
starlette==0.16.0
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.15.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
